+++ src/index.ts
@@ -1,3 +1,4 @@
+export { default as CallbackRegistry } from './CallbackRegistry';
 export { default as ObserverSubject } from './ObserverSubject';
 export { default as DualObserverSubject } from './DualObserverSubject';
 export type { ObserverObject, ObserverFunction } from './DualObserverSubject';